backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
huggingface-hub
sentence-transformers
chroma-db==1.0.15
PyPDF2
pydantic
python-dotenv
aiohttp
python-multipart
tqdm